Completed basic lobby functionality
Play Online sends you to the "lobby" scene. You have the options to create/join a room, enter a room code, and a player name. Room code is generately randomly and player name can be randomized depending on if you input anything. Once the player joins a room, a player list is displayed which updates when players join.